feat(source-core): Moved ability related source files to distinct directory
- To keep things organized with an additional AttributeSet to come
