+++ Source/ProjectLux/Private/Core/AbilitySystem/ProjectLuxCharacterAttributeSet.cpp
@@ -1,7 +1,7 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 
-#include "Core/ProjectLuxCharacterAttributeSet.h"
+#include "Core/AbilitySystem/ProjectLuxCharacterAttributeSet.h"
 #include "GameplayEffectExtension.h"
 #include "Misc/Optional.h"
